Make source location optional
It's not always possible to provide the source location, especially when tracing compiled (Java or .NET) code. However, appmaps lacking it can still be useful. `appmap-java` already omits this information when it's not available.
(I've used 'recommended' instead of 'optional' in line with RFC 2119 -- should we uppercase them and/or use the more familiar MUST/SHOULD/MAY?)

(Aside: should we also make `receiver` and `parameters` optional, seeing as the return value already is?)